signed/unsigned comparison warning fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@20904 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/html/htmlcell.cpp
@@ -372,7 +372,7 @@ void wxHtmlWordCell::Draw(wxDC& dc, int x, int y,
         txt = m_Word.Mid(part1, part2-part1);
         dc.DrawText(txt, ofs + x + m_PosX, y + m_PosY);
 
-        if ( part2 < m_Word.length() )
+        if ( (size_t)part2 < m_Word.length() )
         {
             dc.GetTextExtent(txt, &w, &h);
             ofs += w;
